requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.30.0
bcrypt==4.0.1
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.1.0
Jinja2==3.1.4
kombu==5.5.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.3.7
pluggy==1.6.0
prompt_toolkit==3.0.51
pydantic==2.8.2
pydantic-settings==2.9.1
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==6.2.0
rich==13.7.1
rich-toolkit==0.14.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.37.2
typer==0.12.3
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uvicorn==0.30.3
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
